refactor(location): modularize router with progressive procedures
