+++ app/(main)/locations/[slug]/page.tsx
@@ -9,5 +9,5 @@ export default async function DiscoverLocation({
 }: DiscoverLocationProps) {
 	const { slug } = await params
 	const api = await getAPI()
-	const _location = await api.location.get({ slug })
+	const _location = await api.location.get.core({ slug })
 }
