+++ server/api/routers/location/index.ts
@@ -1,0 +1,8 @@
+import { createTRPCRouter } from '@/server/api/trpc'
+import { locationGetRouter } from './get'
+import { locationListRouter } from './list'
+
+export const locationRouter = createTRPCRouter({
+	list: locationListRouter,
+	get: locationGetRouter,
+})
